+++ ios/democracyclient/AppDelegate.m
@@ -43,7 +43,7 @@
   NSURL *jsCodeLocation;
   
   #ifdef DEBUG
-    jsCodeLocation = [NSURL URLWithString:@"http://192.168.0.24:8081/index.bundle?platform=ios&dev=true"];
+    jsCodeLocation = [NSURL URLWithString:@"http://localhost:8081/index.bundle?platform=ios&dev=true"];
     //jsCodeLocation = [[RCTBundleURLProvider sharedSettings] jsBundleURLForBundleRoot:@"index" fallbackResource:nil];
   #else
     jsCodeLocation = [[NSBundle mainBundle] URLForResource:@"main" withExtension:@"jsbundle"];