+++ CHANGELOG.md
@@ -4,12 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.10.3] - 2018-09-26
+### Added
+- [Statistic] Add new Statistic page
+- [Share] Add share button to detail page
+
+### Fixed
+- [PieCharts] Fix piecharts for ipad 
+
 ## [0.10.2] - 2018-09-19
 ### Added
 - [SMS Verification] Success Screen
 - [Donate] Donate informations
 - [FAQ] New FAQ page
 - [BurgerMenu] New Texts for all Menus
+- [BurgerMenu] Statistic Screen
 
 ### Changed
 - [Filter] changed colors of disabled Sub-Checkboxes [#533](https://github.com/demokratie-live/democracy-client/issues/533)